requirements.txt
# Flask
Flask==0.10.1
Flask-Assets==0.10
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5

# Dicom
pydicom==0.9.9
pillow==2.9.0
numpy==1.9.2

# Tests
pytest==2.7.2
flake8==2.4.1
nose==1.3.7

